+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v5 - 2020-07-23 (not ready yet)
+### Added
+- Odysseyra1n script
+
 ## v4 - 2020-07-21
 ### Added
 - New option: auto recovery (automatically puts the iDevice into recovery mode)